requirements.txt
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
fs==2.4.16
fsspec==2025.3.2
google-auth==2.39.0
greenlet==3.2.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
joblib==1.4.2
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
ply==3.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
simpleeval==1.0.3
simplejson==3.20.1
singer-sdk==0.45.11
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0
